Objects and static variable
Write a program which calculates the number of objects created with the
help of a static variable. Display the result three times such that each time
number of objects created aredifferent. Use a static method to increase the static
variable by one upon every creation of objects and display the result using
another method. Use different class for main() method.