Exceptions in Java
Write an program which handles two types of system generated
exceptions:ArrayIndexOutOfBoundsExceptionandArithmeticException.
When the string “AIOOBE” is passed from command line an object of the type
ArrayIndexOutOfBoundsException is thrown, caught and printed. When the
string “AE” is passed from command line an object of the
type Arithmetic  Exception is thrown, caught and printed. For any other string
passed from command line, the output of the program displayed is “All is well”.
The program also displays the message “Don’t Worry be Happy” as the last line
of the output for every run.