Extra Exceptions in Java
A method in a program manually throws three types of exceptions:
NullPointerException,IllegalAccessExceptionand one user defined exception
named asMyException. Write a program such that when you pass the string
“UE” (meaning unchecked exception) from command-line, the exception
thrown, caught and printed is of the type NullPointerExceptionand when you
pass the string “CE” (meaning checked exception) from command-line, the
exception thrown, caught and printed is of the type IllegalAccessException.
For any other stringthe exception thrown, caught and printed is of the type
MyException.
Take the following as members of the classMyException:
(i) One String variable
(ii) One constructor to initialize the String variable with
“UserDefinedException”.
(iii) toString() method which returns the description
“UserDefinedException.MyException” using the String variable of
this class.